Remove changes which are now obsolete
Some of the changes which were made in the fork are no longer applicable
due to fixes made in more recent versions of proguard. This commit
removes those to bring the fork and main-line source code closer which
will make it easier to make any future merges of upstream releases.